rename confusingly-named ARMCII_Flush_local and associated env var
Flush_local has a meaning in MPI-3 different from ARMCII so I renamed
it Sync_local, which is not identical to any name in MPI but is close
to Sync, which is what it is going to be in the case of MPI-3.
For MPI-2, Sync_local is as before.
The semantics of the two ways to impl this function are different due to
whether or not the window should be locked when the function is called,
hence it might make sense to do something different here.

For example, if we can assume unified, then maybe Sync_local is a no-op.