implement safe in terms of batched to reduce code size
safe is identical to batched with batch_size=1.
we can't just eliminate CONSRV because the safe impl is used
not only in this case but when buffers overlap.

so we added an argument to the batched code that makes it identical
to safe and disabled the safe impl altogether.

some case-switches were trivial condensed to eliminate redundancy
associated with the similarity of PUT and ACC.