implement vector nonblocking calls properly
the vector dispatch code can be generalized to support both B and NB by
adding an argument and conditionalizing the gmr_flush operation
appropriately.

please note that the safe impl is ALWAYS blocking because that seems
like the right thing to do when buffers overlap and no one using CONSRV
should expect good performance anyways.  it seems theoretically possible
to deadlock because of this choice, but that can be addressed when it is
encountered.  it would almost certainly take a deliberately
pathological application to achieve such a result.  for example, if
there are cyclic dependencies and/or synchronization, one might observe
this.  but we know of no such use cases in the wild.