Fix tests (#47)
* tests/mpi/test_mpi_indexed_gets.c: Fix reception buffer initialization

loc_buf was completely uninitialized, while the second and third loop nests
are checking that the values are 1.0 + rank. With luck it would be zero, and
thus the actual - expected > 1e-10 check would pass (since the difference is
then negative). With less luck the check would spuriously fail for the part
that is not overwritten by the MPI_Get.

Signed-off-by: Samuel Thibault <samuel.thibault@ens-lyon.org>

* tests/mpi: Fix comparison of floating-double types

In case the actual value is zero and the expected value is positive, the
check would spuriously succeed.

Signed-off-by: Samuel Thibault <samuel.thibault@ens-lyon.org>

* tests/mpi/test_mpi_indexed_gets.c: Strengthen test

Now that it is fixed, we can make it actually perform strided accesses.

Signed-off-by: Samuel Thibault <samuel.thibault@ens-lyon.org>

---------

Signed-off-by: Samuel Thibault <samuel.thibault@ens-lyon.org>